requirements.txt
numpy==1.13.0
olefile==0.44
Pillow==4.1.1
